Raise on protocols.
The Protocol type(for example Enumerable.t) expands to term/0.
This is not very useful for the user, hence just tell the user to
use the data which implements the protocol as a generator.

This also includes types which will eventually expand to a Protocol,
for example Enum.t is defined as Enumerable.t.