Rewrite recursive types without union.
In the situation you know that the type is `nil | recursive`
you know the leaf upfront.

But other definitions are possible due to `lists = cons (a list) | nil`
Hence when you have a type like `forest = {integer, [forest]}`, you
don't know the leaves upfront. That's why I need to rewrite the whole
type to create the leaf type for this recursive type(In the case of a
forest - that would be `{integer, []}`).