Add recursive types (#17)
* Add the simplest union and recursive types.

Needed to support types such as this one:
  `List a = Nil | Cons a (List a)`

The algorithm to do this is the following:
  * split nodes and leaves
  * create generators out of the leaves
  * use StreamData.tree
  * for the nodes - replace the user type with the leaf

This branch also supports union types such as: Integer or Atom

Done by checking if the nodes are the empty list, if yes -
 return one_of the generated leaves.

* Capture name of the self-referencing type.

Done this by passing in the name along when mapping leafs and the
module.
The module is needed to create a successful generator out of another
user defined type.

* Inline user defined types.

This will ensure that only the user defined type that references the
recursive type will be left.

* Rewrite recursive types without union.

In the situation you know that the type is `nil | recursive`
you know the leaf upfront.

But other definitions are possible due to `lists = cons (a list) | nil`
Hence when you have a type like `forest = {integer, [forest]}`, you
don't know the leaves upfront. That's why I need to rewrite the whole
type to create the leaf type for this recursive type(In the case of a
forest - that would be `{integer, []}`).

* Don't pass unneeded module anymore after inlining.

The module is required only while inlining the user_types.
Afterwards we can discard it and stop passing it around.

For the case of the remote types, they include their own module in the
type signature from the beam code.