[WIP] Add type validations.
Added some passing test cases to showcase what I mean.

Added two examples of how this can be tested.
First one is:
  - generate stream for data
  - generate member function
  - check everything from data return true
  - check everything from term() except manual filter returns false

Second one is using oracles:
  - running example `@type foo :: integer()`
  - create oracle for the member function, in this case this is &is_integer/1
  - generate member function
  - check that for every term() the result of the member and the oracle
  functions is the same

For sake of easier testing, create a public function that uses a type
instead of a function specification. For the function specification
tests only behaviour can be tested.

That way TypesList is reused.