Add support for types with parameters.
This will let people generate stream data generators for generalized
types such as `@type my_type(a) :: list(a)`

The arguments passed in to `from_type` should be a list of the following:
  - any basic type(:integer, :atom, etc.)
  - literal: (1 | 2 | :my_atom | [](empty list) | {} | %{})
  - list: [arguments]
  - tuple: [arguments]
  - map: [{key, value}, {:optional, {key, value}}]
  - user_type: user_type_name(user types are other types defined in the same module)
  - remote_type: {ModuleName, type_name} | {ModuleName, type_name, [arguments]}

Every time a type variable is seen in the type AST, it will be expanded
to a normal type. This also removes the need for arguments to be passed
in.

This also opens up the work for remote types to be done easy.
Remote types already have their arguments expanded, so when working with
remote types with arguments, we can just pass in their arguments to
from_type and don't expand them since that's already done.