Add support for user_types.
Forgot that those are not inlined because the have no arguments.
To make it work, just recursively expand the arguments of the types.

The reason they are not inlined, is because they don't go through the
`inline_user_type` function which usually deals with this. The reason
you cannot call the same function is that you don't know the current
shape of the user type and how the arguments fit, so instead recursively
expand the arguments to call from_type with generators so that it goes
through the inline function.