Add function validation function. (#22)
* Add function validation function.

The function validation API checks whether a function matches it's
defined type specification.

High level overview of how the validation API works:
  1. Read the function spec
  2. Create a generator for the specified arguments
  3. Create a member function for the specified return type
  4. Apply arguments to the function and check that the result belongs
  to the member.
  5. Rescue function failure and return result based on whether the type
  has `:no_return` or `:none`

To do this there is some preparatory work
  - users can name their variables - this is inlined
  - functions with bound variables - also inlined
  - overloaded functions - overloaded functions are tested for every
  overloaded definition there is(the high level overview is  basically
  done for every function definition) - afterwards results are
  aggregated - this could be changed to be better I believe
  - raise when spec/module not found

The current aggregation being done is simple:
  - if all function overloads pass, return {:ok, [list of metadata]}
  - else: {:error, [list of only the failing metadata results]}

* Add documentation.

Removed a wrong a note and added some starting documentation.

* Add another test case.

Tests for overloaded functions with type names.

* Add support for user_types.

Forgot that those are not inlined because the have no arguments.
To make it work, just recursively expand the arguments of the types.

The reason they are not inlined, is because they don't go through the
`inline_user_type` function which usually deals with this. The reason
you cannot call the same function is that you don't know the current
shape of the user type and how the arguments fit, so instead recursively
expand the arguments to call from_type with generators so that it goes
through the inline function.

* Handle no_return functions.

Handling no_return functions should just execute functions
and catch that they error out.

* Assert raise on return.

Do only an assert raise when no_return/none is specified.